+++ src/main/java/com/codigo/monefy/transactions/persistence/entity/Transaction.java
@@ -13,7 +13,7 @@ import java.time.LocalDateTime;
 public class Transaction {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "account_id", nullable = false)
@@ -32,7 +32,7 @@ public class Transaction {
 
     public Transaction() {}
 
-    public Transaction(Long id, Account account, Category category, Double amount, String description, LocalDateTime transactionDate, String transactionType, String notes, String location) {
+    public Transaction(Integer id, Account account, Category category, Double amount, String description, LocalDateTime transactionDate, String transactionType, String notes, String location) {
         this.id = id;
         this.account = account;
         this.category = category;
