+++ src/main/java/com/codigo/monefy/users/persistence/entity/EmailVerificationToken.java
@@ -14,7 +14,7 @@ public class EmailVerificationToken {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id", nullable = false)
@@ -26,7 +26,7 @@ public class EmailVerificationToken {
 
     public EmailVerificationToken() {}
 
-    public EmailVerificationToken(Long id, User user, String token, LocalDateTime expireDate, boolean isRevoked) {
+    public EmailVerificationToken(Integer id, User user, String token, LocalDateTime expireDate, boolean isRevoked) {
         this.id = id;
         this.user = user;
         this.token = token;
